requirements.txt
# This file is used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.26.1=pypi_0
adapters=0.1.1=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
contourpy=1.1.1=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.12.1=pypi_0
fairscale=0.4.4=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=pypi_0
fonttools=4.46.0=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.12.2=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
huggingface-hub=0.19.4=pypi_0
idna=3.6=pypi_0
importlib-resources=6.1.1=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuv=1.44.2=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.4=pypi_0
matplotlib=3.7.4=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=pypi_0
nltk=3.8.1=pypi_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-ml-py=12.535.133=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
nvitop=1.3.2=pypi_0
opencv-python=4.8.1.78=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.12=h7f8727e_0
packaging=23.2=pypi_0
pandas=2.0.3=pypi_0
peft=0.7.1=pypi_0
pillow=10.0.1=py38ha6cbd5a_0
pip=23.3.1=py38h06a4308_0
psutil=5.9.7=pypi_0
pycocoevalcap=1.2=pypi_0
pycocotools=2.0.7=pypi_0
pyparsing=3.1.1=pypi_0
python=3.8.18=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2023.10.3=pypi_0
requests=2.31.0=pypi_0
ruamel-yaml=0.18.5=pypi_0
ruamel-yaml-clib=0.2.8=pypi_0
sacremoses=0.1.1=pypi_0
safetensors=0.4.1=pypi_0
scipy=1.10.1=pypi_0
setuptools=68.2.2=py38h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.12=pypi_0
tbb=2021.8.0=hdb19cb5_0
termcolor=2.4.0=pypi_0
timm=0.4.12=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
torch=2.1.2=pypi_0
torchaudio=0.10.1=py38_cu113
torchvision=0.11.2=py38_cu113
tqdm=4.66.1=pypi_0
transformers=4.26.0=pypi_0
triton=2.1.0=pypi_0
typing_extensions=4.7.1=py38h06a4308_0
tzdata=2023.3=pypi_0
urllib3=2.1.0=pypi_0
wheel=0.41.2=py38h06a4308_0
xz=5.4.5=h5eee18b_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
